+++ frontend/src/components/ConfirmAccount/ConfirmAccount.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import { useParams, Link } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import { validateUser } from "../redux/actions/actionUser.js";
+import { validateUser } from "../../redux/actions/actionUser.js";
 
 export default function ConfirmarCuenta() {
   const dispatch = useDispatch();
