+++ frontend/src/components/NewPassword/NewPassword.jsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useParams } from "react-router";
 import { Link } from "react-router-dom";
-import { resetPassword, setStateEmail } from "../redux/actions/actionUser.js";
+import { resetPassword, setStateEmail } from "../../redux/actions/actionUser.js";
 
-import validatePassword from "../middleware/validarPassword";
+import validatePassword from "../../middleware/validarPassword";
 
 export default function NewPassword() {
   const params = useParams();
